requirements.txt
langchain==0.0.167
openai==0.27.6
pydantic==1.8.2
python-dotenv==0.19.1
rich==13.3.1
pyyaml==6.0
streamlit==1.19.0
altair<5